requirements.txt
numpy
pandas
scipy
PyYAML
netCDF4
xarray
matplotlib
